requirements.txt
# Core
requests
numpy
pillow
tqdm

# TCGA WSI tiling
openslide-python

# Evaluation
scikit-learn

# MedGemma (multimodal Gemma 3)
# Gemma 3 support starts from transformers 4.50.0
transformers>=4.50.0
accelerate

# Install torch separately depending on your CUDA setup
# torch
